perf: improve rust hex (de-)serialization (#3915)
While these were not very expensive, they are fairly commonly called and
improving the performance here was a nice, quick win. On top of that, we
removed the heap allocation for `hex_to_bytes` which should improve
performance even further on lower-end machines.
